sandbox/cover-letter-generator/requirements.txt
# Core dependencies
streamlit>=1.31.0
deepseek>=0.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
